requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
black==23.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==41.0.1
Deprecated==1.2.14
discord==2.2.2
discord.py==2.2.2
fastapi==0.95.0
frozenlist==1.3.3
future==0.18.3
h11==0.14.0
idna==3.4
ifaddr==0.2.0
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.1
platformdirs==3.2.0
pycparser==2.21
pydantic==1.10.7
PyGithub==1.58.2
PyJWT==2.7.0
PyNaCl==1.5.0
pypiwin32==223
python-dotenv==1.0.0
pytz==2023.3
pywin32==306
requests==2.31.0
sniffio==1.3.0
starlette==0.26.1
typing_extensions==4.5.0
urllib3==2.0.3
uvicorn==0.21.1
wrapt==1.15.0
yarl==1.8.2
yeelight==0.7.10
